backend/requirements.txt
# Python 3.10+ required
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlink>=7.5.0
apscheduler==3.10.4
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
aiosqlite==0.19.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
factory-boy==3.3.0
httpx==0.25.2
aiohttp==3.9.1
psutil==5.9.6
matplotlib>=3.5.0
seaborn>=0.11.0
